Roadmap and tests

+++ src/FolkerKinzel.Strings/FileInfoExtension.cs
@@ -1,8 +1,11 @@
 ﻿namespace FolkerKinzel.Strings;
 
+/// <summary>
+/// Erweiterungsmethoden für die <see cref="FileInfo"/>-Klasse.
+/// </summary>
 public static class FileInfoExtension
 {
-    public static bool IsUtf8(this FileInfo fileInfo, long count = -1)
+    public static bool IsUtf8(this FileInfo fileInfo, int count = -1)
     {
         if (fileInfo == null)
         {
@@ -13,7 +16,7 @@ public static class FileInfoExtension
         return stream.IsUtf8Internal(count, false);
     }
 
-    public static bool IsValidUtf8(this FileInfo fileInfo, long count = -1)
+    public static bool IsValidUtf8(this FileInfo fileInfo, int count = -1)
     {
         if (fileInfo == null)
         {
